OCPBUGS-81319: Move coFetch to core-api only
Previously, plugins importing the `console-fetch-utils` module would face module resolution errors because a helper function erroneously used a dynamic `import` instead of a `require`, which triggers webpack import validation (which of course does not work in a dynamic plugin context)

Exposed HTTPError as official API because it's thrown by consoleFetch and would be needed for `instanceof` checks.

There are still type errors due to use of window.SERVER_FLAGS, so it's probably better to use dynamic require for the whole module and move the code to console-shared

getCSRFToken is needed by monitoring-plugin, so it's moved to the internal package.
