+++ frontend/packages/console-dynamic-plugin-sdk/src/api/core-api.ts
@@ -39,6 +39,9 @@ import {
   UseResolvedExtensions,
   UseUserPreference,
   VirtualizedTableFC,
+  ConsoleFetch,
+  ConsoleFetchJSON,
+  ConsoleFetchText
 } from '../extensions/console-types';
 import { StatusPopupSectionProps, StatusPopupItemProps } from '../extensions/dashboard-types';
 
@@ -398,7 +401,49 @@ export {
   useK8sWatchResources,
 } from '../utils/k8s/hooks';
 
-export { consoleFetch, consoleFetchJSON, consoleFetchText } from '../utils/fetch';
+// Error thrown by consoleFetch and consoleFetchJSON. Exposed for `instanceOf` checks
+export { HttpError } from '../utils/error/http-error';
+
+/**
+ * A custom wrapper around `fetch` that adds console-specific headers and allows for retries and timeouts.
+ * It also validates the response status code and throws an appropriate error or logs out the user if required.
+ * @param url - The URL to fetch
+ * @param options - The options to pass to fetch
+ * @param timeout - The timeout in milliseconds
+ * @returns A promise that resolves to the response.
+ * @throws {@link HttpError} when the response status code indicates an error
+ */
+export const consoleFetch: ConsoleFetch = require('@console/shared/src/utils/console-fetch')
+  .coFetch;
+
+/**
+ * A custom wrapper around `fetch` that adds console-specific headers and allows for retries and timeouts.
+ * It also validates the response status code and throws an appropriate error or logs out the user if required.
+ * It returns the response as a JSON object.
+ * Uses consoleFetch internally.
+ * @param url The URL to fetch
+ * @param method  The HTTP method to use. Defaults to GET
+ * @param options The options to pass to fetch
+ * @param timeout The timeout in milliseconds
+ * @returns A promise that resolves to the response as text or JSON object.
+ * @throws {@link HttpError} when the response status code indicates an error
+ */
+export const consoleFetchJSON: ConsoleFetchJSON = require('@console/shared/src/utils/console-fetch')
+  .coFetchJSON;
+
+/**
+ * A custom wrapper around `fetch` that adds console-specific headers and allows for retries and timeouts.
+ * It also validates the response status code and throws an appropriate error or logs out the user if required.
+ * It returns the response as a text.
+ * Uses `consoleFetch` internally.
+ * @param url The URL to fetch
+ * @param options The options to pass to fetch
+ * @param timeout The timeout in milliseconds
+ * @returns A promise that resolves to the response as text or JSON object.
+ * @throws {@link HttpError} when the response status code indicates an error
+ */
+export const consoleFetchText: ConsoleFetchText = require('@console/shared/src/utils/console-fetch')
+  .coFetchText;
 
 // Expose K8s CRUD utilities as below
 export {
