+++ frontend/packages/console-dynamic-plugin-sdk/src/api/internal-types.ts
@@ -399,3 +399,5 @@ export type SwaggerDefinitions = {
 };
 
 export type DefinitionFor = (model: K8sModel) => SwaggerDefinition;
+
+export type GetCSRFToken = () => string;
