--- frontend/packages/console-dynamic-plugin-sdk/src/utils/fetch/index.ts
@@ -1,2 +1,0 @@
-export { consoleFetch, consoleFetchJSON, consoleFetchText } from './console-fetch';
-export { getConsoleRequestHeaders, normalizeConsoleHeaders } from './console-fetch-utils';
