+++ frontend/packages/console-dynamic-plugin-sdk/src/utils/k8s/k8s-resource.ts
@@ -1,8 +1,8 @@
 import * as _ from 'lodash';
 import type { K8sModel } from '../../api/common-types';
+import { coFetchJSON } from '@console/shared/src/utils/console-fetch';
 import type { Options } from '../../api/internal-types';
 import type { K8sResourceCommon, Patch, QueryParams } from '../../extensions/console-types';
-import { consoleFetchJSON as coFetchJSON } from '../fetch';
 import { selectorToString, resourceURL } from './k8s-utils';
 
 type BaseOptions = {
