+++ frontend/public/components/utils/__tests__/download-button.spec.tsx
@@ -2,7 +2,7 @@ import { render, screen, fireEvent, waitFor, act } from '@testing-library/react'
 import * as fileSaver from 'file-saver';
 
 import { DownloadButton } from '../../../components/utils/download-button';
-import * as coFetchModule from '@console/dynamic-plugin-sdk/src/utils/fetch/console-fetch';
+import * as coFetchModule from '@console/shared/src/utils/console-fetch';
 
 // Mock file-saver
 jest.mock('file-saver', () => ({
@@ -10,13 +10,13 @@ jest.mock('file-saver', () => ({
 }));
 
 // Mock console-fetch
-jest.mock('@console/dynamic-plugin-sdk/src/utils/fetch/console-fetch', () => ({
-  consoleFetch: jest.fn(),
+jest.mock('@console/shared/src/utils/console-fetch', () => ({
+  coFetch: jest.fn(),
 }));
 
 describe('DownloadButton', () => {
   const url = 'http://google.com';
-  const mockConsoleFetch = coFetchModule.consoleFetch as jest.Mock;
+  const mockConsoleFetch = coFetchModule.coFetch as jest.Mock;
   const mockSaveAs = fileSaver.saveAs as jest.Mock;
 
   beforeEach(() => {
